# loan-service/docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Loan Service: registers loans, returns, and validates with Catalog Service that the book exists and has stock. Exposes HTTP and gRPC; Catalog Service consumes HTTP by default and gRPC as a bonus.\n\nReverse channel: Loan Service calls Catalog Service via HTTP (`/books/:id`) or gRPC (`BookExists`) to verify availability before registering a loan.",
        "title": "Loan Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/availability/{bookId}": {
            "get": {
                "description": "Loan Service calls Catalog Service to confirm the book still exists and has stock. This is the reverse channel used by the saga.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "Check book availability (reverse channel)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book UUID",
                        "name": "bookId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.availabilityResponse"
                        }
                    },
                    "400": {
                        "description": "invalid uuid",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "book not found in catalog",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "catalog service unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "description": "Returns 200 with db=up when the database is reachable; 503 with db=down otherwise. Fast and dependency-free for orchestrator healthchecks.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Liveness + DB ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/loans": {
            "get": {
                "description": "Returns a paginated list of loans. Supports filtering by user_id, book_id, status, and date range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "List loans (paginated, filterable)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by user UUID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by book UUID",
                        "name": "book_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "active",
                            "returned"
                        ],
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Borrowed at \u003e= from (RFC3339)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Borrowed at \u003c= to (RFC3339)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "1-based page index",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page (1-100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoanListResponse"
                        }
                    },
                    "400": {
                        "description": "invalid query",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Validates with Catalog Service that the book exists and has stock, then persists the loan. Idempotent if `idempotency_key` is supplied.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Register a loan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional dedup key (also accepted in body)",
                        "name": "X-Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Loan request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateLoanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.LoanResponse"
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "book not found in catalog",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "insufficient stock or duplicate idempotency key",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "catalog service unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/loans/{id}/return": {
            "post": {
                "description": "Marks the loan as returned and updates the timestamp. Idempotent: returning an already-returned loan is a no-op.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Register a return",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoanResponse"
                        }
                    },
                    "400": {
                        "description": "invalid uuid",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "loan not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateLoanRequest": {
            "type": "object",
            "required": [
                "book_id",
                "user_id"
            ],
            "properties": {
                "book_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "978-0-13-468599-1"
                },
                "idempotency_key": {
                    "type": "string",
                    "example": "loan-attempt-2025-01-15-abc123"
                },
                "user_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "3e1ae706-5ca9-4aeb-aa60-c5b7f942969d"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
        "dto.LoanListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.LoanResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.LoanResponse": {
            "type": "object",
            "properties": {
                "book_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "978-0-13-468599-1"
                },
                "borrowed_at": {
                    "type": "string",
                    "example": "2025-01-15T10:00:00Z"
                },
                "id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "3e1ae706-5ca9-4aeb-aa60-c5b7f942969d"
                },
                "returned_at": {
                    "type": "string",
                    "example": "2025-02-15T10:00:00Z"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "returned"
                    ],
                    "example": "active"
                },
                "user_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "3e1ae706-5ca9-4aeb-aa60-c5b7f942969d"
                }
            }
        },
        "http.availabilityResponse": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "availableStock": {
                    "type": "integer"
                },
                "exists": {
                    "type": "boolean"
                },
                "totalStock": {
                    "type": "integer"
                }
            }
        }
    }
}